Point to archive.org mirror of the radiance data
Data required for the build used to be at http://www.graphics.cornell.edu/resources/clearsky/data/2013-05-27/RADIANCE/ but is no longer there. A mirror of the data is provided at https://www.yiningkarlli.com/projects/skydomecompare/data.html, but it is non-trivial make this work with the Makefile. Changed link to point to the archived version of the data from archive.org.